+++ tests/Integration/Captcha/Shop/UserComponentTest.php
@@ -45,6 +45,14 @@ class UserComponentTest extends IntegrationTestCase
         Registry::getSession()->setVariable('captcha_expiration', time() + 60);
     }
 
+    public function tearDown(): void
+    {
+        $moduleSettings = ContainerFacade::get(ModuleSettingsServiceInterface::class);
+        $moduleSettings->saveIsCaptchaEnabled(false);
+
+        parent::tearDown();
+    }
+
     public function testLoginWithInvalidCaptchaReturnsUserAndShowsError(): void
     {
         $this->requestMock
@@ -114,7 +122,7 @@ class UserComponentTest extends IntegrationTestCase
             ->willReturn('invalid_captcha');
 
         $this->utilsViewMock
-            ->expects($this->atLeastOnce())
+            ->expects($this->once())
             ->method('addErrorToDisplay')
             ->with('ERROR_INVALID_CAPTCHA');
 
@@ -122,5 +130,126 @@ class UserComponentTest extends IntegrationTestCase
         $result = $subject->createUser();
 
         $this->assertFalse($result);
+    }
+
+    public function testLoginWithValidCaptchaDoesNotShowCaptchaError(): void
+    {
+        $this->requestMock
+            ->method('getRequestParameter')
+            ->willReturnCallback(function ($param) {
+                if ($param === 'captcha') {
+                    return 'valid_captcha';
+                }
+                return '';
+            });
+
+        $this->requestMock
+            ->method('getRequestEscapedParameter')
+            ->willReturn('');
+
+        $this->utilsViewMock
+            ->expects($this->any())
+            ->method('addErrorToDisplay')
+            ->willReturnCallback(function ($message) {
+                $this->assertNotEquals('ERROR_INVALID_CAPTCHA', $message);
+                $this->assertNotEquals('ERROR_EMPTY_CAPTCHA', $message);
+                $this->assertNotEquals('FORM_VALIDATION_FAILED', $message);
+            });
+
+        $subject = oxNew(UserComponent::class);
+        $subject->login();
+    }
+
+    public function testCreateUserWithValidCaptchaDoesNotShowCaptchaError(): void
+    {
+        $this->requestMock
+            ->method('getRequestParameter')
+            ->willReturnCallback(function ($param) {
+                if ($param === 'captcha') {
+                    return 'valid_captcha';
+                }
+                return '';
+            });
+
+        $this->utilsViewMock
+            ->expects($this->any())
+            ->method('addErrorToDisplay')
+            ->willReturnCallback(function ($message) {
+                $this->assertNotEquals('ERROR_INVALID_CAPTCHA', $message);
+                $this->assertNotEquals('ERROR_EMPTY_CAPTCHA', $message);
+                $this->assertNotEquals('FORM_VALIDATION_FAILED', $message);
+            });
+
+        $subject = oxNew(UserComponent::class);
+        $subject->createUser();
+    }
+
+    public function testCreateUserWithInvalidHoneyPotReturnsFalseAndShowsError(): void
+    {
+        $this->requestMock
+            ->method('getRequestParameter')
+            ->willReturnCallback(function ($param) {
+                if ($param === 'captcha') {
+                    return 'valid_captcha';
+                }
+                if ($param === 'lastname_confirm') {
+                    return 'some-text';
+                }
+                return '';
+            });
+
+        $this->utilsViewMock
+            ->expects($this->once())
+            ->method('addErrorToDisplay')
+            ->with('FORM_VALIDATION_FAILED');
+
+        $subject = oxNew(UserComponent::class);
+        $result = $subject->createUser();
+
+        $this->assertFalse($result);
+    }
+
+    public function testLoginWithCaptchaDisabledDoesNotValidateCaptcha(): void
+    {
+        $moduleSettings = ContainerFacade::get(ModuleSettingsServiceInterface::class);
+        $moduleSettings->saveIsCaptchaEnabled(false);
+
+        $this->requestMock
+            ->method('getRequestParameter')
+            ->willReturn('');
+
+        $this->requestMock
+            ->method('getRequestEscapedParameter')
+            ->willReturn('');
+
+        $this->utilsViewMock
+            ->expects($this->any())
+            ->method('addErrorToDisplay')
+            ->willReturnCallback(function ($message) {
+                $this->assertNotEquals('ERROR_INVALID_CAPTCHA', $message);
+                $this->assertNotEquals('ERROR_EMPTY_CAPTCHA', $message);
+                $this->assertNotEquals('FORM_VALIDATION_FAILED', $message);
+            });
+
+        $subject = oxNew(UserComponent::class);
+        $subject->login();
+    }
+
+    public function testCreateUserWithCaptchaDisabledDoesNotValidateCaptcha(): void
+    {
+        $moduleSettings = ContainerFacade::get(ModuleSettingsServiceInterface::class);
+        $moduleSettings->saveIsCaptchaEnabled(false);
+
+        $this->utilsViewMock
+            ->expects($this->any())
+            ->method('addErrorToDisplay')
+            ->willReturnCallback(function ($message) {
+                $this->assertNotEquals('ERROR_INVALID_CAPTCHA', $message);
+                $this->assertNotEquals('ERROR_EMPTY_CAPTCHA', $message);
+                $this->assertNotEquals('FORM_VALIDATION_FAILED', $message);
+            });
+
+        $subject = oxNew(UserComponent::class);
+        $subject->createUser();
     }
 }