+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.0.0] - Unreleased
 
 ### Fixed
-- Compatibility with api login
+- Captcha validation moved from User model to UserComponent to prevent API login failures
 
 ## [2.1.0] - 2026-01-14